refactor: use a traversal method
